Conflicts:
	src/librtlsdr.c
	src/tuner_r82xx.c

+++ src/librtlsdr.c
@@ -705,7 +705,7 @@ static int rtl2832_set_if_freq(rtlsdr_dev_t *dev, uint32_t freq, uint32_t *freq_
 	if (rtlsdr_get_xtal_freq(dev, &rtl_xtal, NULL))
 		return -2;
 
-	if_freq = ((rtl_xtal/2 + (uint64_t)freq * TWO_POW(22)) / rtl_xtal) * (-1);
+	if_freq = (int32_t)(((rtl_xtal/2 + (uint64_t)freq * TWO_POW(22)) / rtl_xtal) * (-1));
 	if (if_freq <= -0x200000) {
 		/* fprintf(stderr, "rtl2832_set_if_freq(): %u Hz out of range for downconverter (divisor would be %x)\n", freq, if_freq); */
 		return -2;
@@ -727,7 +727,7 @@ int rtlsdr_set_sample_freq_correction(rtlsdr_dev_t *dev, int ppm)
 {
 	int r = 0;
 	uint8_t tmp;
-	int16_t offs = ppm * (-1) * TWO_POW(24) / 1000000;
+	int16_t offs = (int16_t)(ppm * -1 * TWO_POW(24) / 1000000);
 	rtlsdr_set_i2c_repeater(dev, 0);
 
 	tmp = offs & 0xff;
@@ -1137,7 +1137,7 @@ int rtlsdr_set_sample_rate(rtlsdr_dev_t *dev, uint32_t samp_rate)
 		return -EINVAL;
 	}
 
-	rsamp_ratio = (dev->rtl_xtal * TWO_POW(22)) / samp_rate;
+	rsamp_ratio = (uint32_t)((dev->rtl_xtal * TWO_POW(22)) / samp_rate);
 	rsamp_ratio &= 0x0ffffffc;
 
 	real_rsamp_ratio = rsamp_ratio | ((rsamp_ratio & 0x08000000) << 1);
@@ -1858,7 +1858,7 @@ int rtlsdr_read_async(rtlsdr_dev_t *dev, rtlsdr_read_async_cb_t cb, void *ctx,
 
 		r = libusb_submit_transfer(dev->xfer[i]);
 		if (r < 0) {
-			fprintf(stderr, "Failed to submit transfer %i!\n", i);
+			fprintf(stderr, "Failed to submit transfer %u!\n", i);
 			dev->async_status = RTLSDR_CANCELING;
 			break;
 		}
